Move modals into its own file
