+++ src/modules/auto_response/shared_cache.py
@@ -1,0 +1,3 @@
+from resources.models.autoresponse import AutoResponse
+
+stored_trigger_map: dict[str, AutoResponse] = dict()
